Add income/expenditure
I did not think to create a Git Hub which is why there is so much work being pushed in one go.
I have created the interface and the implementation for add income/expenditure. It is relatively robust in its current state but has not been formally tested yet.
I also added my own basic exception class for an Invalid input form the user as I might deicide to mess with custom exceptions later into the project.